refactored Questioneer to separate files, for clean code
